Only generate vApp config when targeting containerlinux

+++ pkg/cloudprovider/provider/vsphere/provider.go
@@ -188,7 +188,7 @@ func machineInvalidConfigurationTerminalError(err error) error {
 }
 
 func (p *provider) Create(machine *v1alpha1.Machine, userdata string) (instance.Instance, error) {
-	config, _, err := p.getConfig(machine.Spec.ProviderConfig)
+	config, pc, err := p.getConfig(machine.Spec.ProviderConfig)
 	if err != nil {
 		return nil, fmt.Errorf("failed to parse config: %v", err)
 	}
@@ -198,6 +198,11 @@ func (p *provider) Create(machine *v1alpha1.Machine, userdata string) (instance.
 		return nil, fmt.Errorf("failed to get vsphere client: '%v'", err)
 	}
 
+	var containerLinuxUserdata string
+	if pc.OperatingSystem == providerconfig.OperatingSystemCoreos {
+		containerLinuxUserdata = userdata
+	}
+
 	_, err = CreateLinkClonedVm(machine.Spec.Name,
 		config.TemplateVMName,
 		config.Datacenter,
@@ -205,7 +210,7 @@ func (p *provider) Create(machine *v1alpha1.Machine, userdata string) (instance.
 		config.CPUs,
 		config.MemoryMB,
 		client,
-		userdata)
+		containerLinuxUserdata)
 	if err != nil {
 		return nil, machineInvalidConfigurationTerminalError(fmt.Errorf("failed to create linked vm: '%v'", err))
 	}
@@ -219,21 +224,23 @@ func (p *provider) Create(machine *v1alpha1.Machine, userdata string) (instance.
 		return nil, fmt.Errorf("failed to get virtual machine object: %v", err)
 	}
 
-	localUserdataIsoFilePath, err := generateLocalUserdataIso(userdata, machine.Spec.Name)
-	if err != nil {
-		return nil, err
-	}
-
-	defer func() {
-		err := os.Remove(localUserdataIsoFilePath)
+	if pc.OperatingSystem != providerconfig.OperatingSystemCoreos {
+		localUserdataIsoFilePath, err := generateLocalUserdataIso(userdata, machine.Spec.Name)
 		if err != nil {
-			utilruntime.HandleError(fmt.Errorf("failed to clean up local userdata iso file at %s: %v", localUserdataIsoFilePath, err))
+			return nil, err
 		}
-	}()
 
-	err = uploadAndAttachISO(finder, virtualMachine, localUserdataIsoFilePath, config.Datastore, client)
-	if err != nil {
-		return nil, machineInvalidConfigurationTerminalError(fmt.Errorf("failed to upload and attach userdata iso: %v", err))
+		defer func() {
+			err := os.Remove(localUserdataIsoFilePath)
+			if err != nil {
+				utilruntime.HandleError(fmt.Errorf("failed to clean up local userdata iso file at %s: %v", localUserdataIsoFilePath, err))
+			}
+		}()
+
+		err = uploadAndAttachISO(finder, virtualMachine, localUserdataIsoFilePath, config.Datastore, client)
+		if err != nil {
+			return nil, machineInvalidConfigurationTerminalError(fmt.Errorf("failed to upload and attach userdata iso: %v", err))
+		}
 	}
 
 	// Ubuntu wont boot with attached floppy device, because it tries to write to it
